invoice_triage_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-triage-env"
version = "0.1.0"
description = "OpenEnv environment for AI agent invoice processing and triage"
license = "BSD-3-Clause"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
server = [
    "uvicorn>=0.30.0",
    "fastapi>=0.110.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
invoice-triage-server = "invoice_triage_env.server.app:main"

[tool.setuptools.packages.find]
include = ["invoice_triage_env*"]
